requirements.txt
python-telegram-bot==20.*
yt-dlp
python-dotenv
aiohttp
aiofiles